Add FusionCacheInternalStrings + tests
