+++ tests/ZiggyCreatures.FusionCache.Tests/AutoRecoveryTests_Async.cs
@@ -1,7 +1,6 @@
 ﻿using FusionCacheTests.Stuff;
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Caching.Memory;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Xunit;
 using ZiggyCreatures.Caching.Fusion;
